test(mfa): extract transition-coverage analysis into transitionCoverage.ts
flowPaths held both the paths the walk executes and the analysis that
measures those paths against the machine's adjacency map, so answering
where the UI paths are defined meant reading past the coverage half.
The analysis moves to transitionCoverage.ts, and the shared
getStateChangingEdges names the adjacency computation the two adjacency
consumers previously duplicated.
